Changes to support replication and alwayson
added config options to allow SQL Trace
added config options for multiple Trace and XEvent templates
added scripts to gather AlwaysOn, QueryStore information
fixed bugs where sql port detection had issues when MSDTC is enabled

+++ LinuxPSSDiag/stop_collector.sh
@@ -7,21 +7,58 @@ source ./pssdiag_support_functions.sh
 
 sql_collect_config()
 {
-	echo "	Collecting SQL Configuration Snapshot script at Shutdown..."
+	echo "	Collecting SQL Configuration Snapshot at Shutdown..."
         /opt/mssql-tools/bin/sqlcmd -S$SQL_SERVER_NAME -U$sqluser -P$pass -i"SQL_Configuration.sql" -o"$outputdir/${1}_SQL_Configuration_Shutdown.out"
 }
 
 sql_stop_xevent()
 {
+if [[ $COLLECT_EXTENDED_EVENTS == [Yy][eE][sS]  ]]; then
 	echo "	Stopping XE Collection if started..."
         /opt/mssql-tools/bin/sqlcmd -S$SQL_SERVER_NAME -U$sqluser -P$pass -i"pssdiag_xevent_stop.sql" -o"$outputdir/${1}_Stop_XECollection.out"
+fi
+}
+
+sql_stop_trace()
+{
+if [[ $COLLECT_SQL_TRACE == [Yy][eE][sS]  ]]; then
+        echo "	Stopping SQL Trace Collection if started..."
+        /opt/mssql-tools/bin/sqlcmd -S$SQL_SERVER_NAME -U$sqluser -P$pass -i"pssdiag_trace_stop.sql" -o"$outputdir/${1}_Stop_TraceCollection.out"
+fi
 }
 
 sql_collect_xevent()
 {
+if [[ $COLLECT_EXTENDED_EVENTS == [Yy][eE][sS]  ]]; then
 	docker exec $1 sh -c "cd /var/opt/mssql/log  && tar cf /tmp/sql_xevent.tar *pssdiag_xevent*.xel "
         docker cp $1:/tmp/sql_xevent.tar ${outputdir}/${2}_sql_xevent.tar
         docker exec $1 sh -c "rm -f /tmp/sql_xevent.tar"
+fi
+}
+
+sql_collect_trace()
+{
+if [[ $COLLECT_SQL_TRACE == [Yy][eE][sS]  ]]; then
+        docker exec $1 sh -c "cd /var/opt/mssql/log  && tar cf /tmp/sql_trace.tar *pssdiag_trace*.trc "
+        docker cp $1:/tmp/sql_trace.tar ${outputdir}/${2}_sql_trace.tar
+        docker exec $1 sh -c "rm -f /tmp/sql_trace.tar"
+fi
+}
+
+sql_collect_alwayson()
+{
+if [[ $COLLECT_SQL_HA_LOGS == [Yy][eE][sS]  ]]; then
+        echo "	Collecting SQL AlwaysOn configuration at Shutdown..."
+        /opt/mssql-tools/bin/sqlcmd -S$SQL_SERVER_NAME -U$sqluser -P$pass -i"SQL_AlwaysOnDiagScript.sql" -o"$outputdir/${1}_SQL_AlwaysOnDiag_Shutdown.out"
+fi
+}
+
+sql_collect_querystore()
+{
+if [[ $COLLECT_QUERY_STORE == [Yy][eE][sS]  ]]; then
+        echo "	Collecting SQL Query Store information at Shutdown..."
+        /opt/mssql-tools/bin/sqlcmd -S$SQL_SERVER_NAME -U$sqluser -P$pass -i"SQL_QueryStore.sql" -o"$outputdir/${1}_SQL_QueryStore_Shutdown.out"
+fi
 }
 
 # end of function definitions
@@ -64,6 +101,11 @@ COLLECT_SQL_DUMPS=${COLLECT_SQL_DUMPS:-"NO"}
 COLLECT_SQL_LOGS=${COLLECT_SQL_LOGS:-"NO"}
 COLLECT_OS_LOGS=${COLLECT_OS_LOGS:-"NO"}
 COLLECT_OS_CONFIG=${COLLECT_OS_CONFIG:-"NO"}
+COLLECT_EXTENDED_EVENTS=${COLLECT_EXTENDED_EVENTS:-"NO"}
+COLLECT_SQL_TRACE=${COLLECT_SQL_TRACE:-"NO"}
+COLLECT_QUERY_STORE=${COLLECT_QUERY_STORE:-"NO"}
+COLLECT_SQL_HA_LOGS=${COLLECT_SQL_HA_LOGS:-"NO"}
+
 
 #collect basic machine configuration
 if [[ $COLLECT_OS_CONFIG == "YES" ]] ; then
@@ -97,9 +139,11 @@ if [[ "$COLLECT_HOST_SQL_INSTANCE" == "YES" ]];then
         else
 		sql_collect_config "${HOSTNAME}"
 		sql_stop_xevent "${HOSTNAME}"
-		 #chown only if pattern exists.
-                 stat -t -- $output/*.xel >/dev/null 2>&1 && chown $USER: $outputdir/*.xel
-		
+		sql_stop_trace "${HOSTNAME}"
+		#chown only if pattern exists.
+                stat -t -- $output/*.xel >/dev/null 2>&1 && chown $USER: $outputdir/*.xel
+		sql_collect_alwayson "${HOSTNAME}"
+		sql_collect_querystore "${HOSTNAME}"
 	fi
 fi
 
@@ -119,6 +163,10 @@ if [[ "$COLLECT_CONTAINER" != "NO" ]]; then
 			sql_collect_config "${dockername}"
 			sql_stop_xevent "${dockername}"
 			sql_collect_xevent "${dockerid}" "${dockername}"
+			sql_stop_trace "${dockername}"
+			sql_collect_trace "${dockerid}" "${dockername}"
+			sql_collect_alwayson "${dockername}"
+			sql_collect_querystore "${dockername}"
                 fi
         # we finished processing the requested container
         else
@@ -137,6 +185,10 @@ if [[ "$COLLECT_CONTAINER" != "NO" ]]; then
 				sql_collect_config "${dockername}"
 				sql_stop_xevent "${dockername}"
 				sql_collect_xevent "${dockerid}" "${dockername}"
+				sql_stop_trace "${dockername}"
+				sql_collect_trace "${dockerid}" "${dockername}"
+        	                sql_collect_alwayson "${dockername}"
+	                        sql_collect_querystore "${dockername}"
                         fi
                 done;
         # we finished processing all the container
